[#148] Removed unused files (routes/questions.js) and renamed other files (question-prompt.js) to more closely match naming convention
